+++ apps/formio-server/src/models/models.js
@@ -2,6 +2,7 @@
 
 module.exports = function(router) {
   return {
-    project: require('./Project')(router)
+    project: require('./Project')(router),
+    tag: require('./Tag')(router)
   };
 };